Drop redundant .to() casts in fused tree-decode buffer prep
run_fused_tree_decode_cudagraph was calling .to(torch.int32) / .to(torch.int64)
before assigning into the graph-bound tensors. That pattern allocates a
temporary and then copies — slice-assign already performs the dtype cast
in a single kernel.

Tiny win (~0.05 ms at K=15 on the 8B eagle bench) but truly free: fewer
kernels, fewer transient allocations. A follow-up that plumbed optional
out-buffer arguments through _compute_step_positions_and_slot_maps was
tested and reverted — PyTorch's caching allocator made the eliminated
allocations nearly free, so no measurable production win despite the
extra code.

+++ ssd/engine/helpers/cudagraph_helpers.py
@@ -993,12 +993,15 @@ def run_fused_tree_decode_cudagraph(
     # Per-step metadata: copy the first orig_flat entries per step; padded positions
     # default to zeros which point at slot 0.  Attention outputs at padded positions
     # are discarded; we only read [:orig_flat] after replay.
-    gv["step_slot_maps"][:K, :orig_flat] = step_slot_maps.to(torch.int32)
-    gv["step_rope_positions"][:K, :orig_flat] = step_rope_positions.to(torch.int64)
-    gv["step_context_lens"][:K, :orig_B] = step_context_lens.to(torch.int32)
+    #
+    # PERFORMANCE: No explicit .to(dtype) — slice-assign casts in a single kernel,
+    # whereas .to() allocates a temp and then copies (two kernels).
+    gv["step_slot_maps"][:K, :orig_flat] = step_slot_maps
+    gv["step_rope_positions"][:K, :orig_flat] = step_rope_positions
+    gv["step_context_lens"][:K, :orig_B] = step_context_lens
     if bucket_bs > orig_B:
         # Ghost seqs: repeat last real seq's context_len + block_table to keep FA4 happy.
-        gv["step_context_lens"][:K, orig_B:bucket_bs] = step_context_lens[:, -1:].to(torch.int32)
+        gv["step_context_lens"][:K, orig_B:bucket_bs] = step_context_lens[:, -1:]
         gv["block_tables"][orig_B:bucket_bs, :dbt.shape[1]] = dbt[-1:].expand(bucket_bs - orig_B, -1)
     gv["block_tables"][:orig_B, :dbt.shape[1]] = dbt
